deep conv resnet at work

+++ conv_autoencoder.py
@@ -2,12 +2,14 @@ import xarray as xr
 import numpy as np
 import matplotlib.pyplot as plt
 import xesmf as xe
+import os
 import time
 from xmovie import Movie
 from multiprocess import Pool
 from dask.diagnostics import ProgressBar
 from importlib import reload
 from ESN.ESN import ESN
+from datetime import datetime
 
 import torch
 import torch.nn as nn
@@ -19,13 +21,13 @@ import xbatcher
 # Load training and test data
 
 HR_data_file = ('cmems_mod_nws_phy_anfc_0.027deg-2D_PT15M-i_'
-                'uo-vo_4.23E-7.78E_56.81N-58.70N_2023-01-01-2023-02-01.nc')
+                'uo-vo_4.23E-7.78E_56.81N-58.70N_2023-01-01-2023-05-01.nc')
 
 HR_bathy_file = ('cmems_mod_nws_phy_anfc_0.027deg-3D_'
                  'static_multi-vars_4.23E-7.78E_56.81N-58.70N_0.49-643.57m.nc')
 
 LR_data_file = ('cmems_mod_nws_phy-uv_my_7km-2D_PT1H-i_'
-                'uo-vo_4.22E-7.78E_56.80N-58.67N_2023-01-01-2023-02-01.nc')
+                'uo-vo_4.22E-7.78E_56.80N-58.67N_2023-01-01-2023-05-01.nc')
 
 bt_HR = xr.open_dataset(HR_bathy_file)
 
@@ -99,17 +101,41 @@ def normalize(x, minval=None, maxmin=None):
         minval = np.nanmin(x)
     if maxmin is None:
         maxmin = np.nanmax(x) - minval
-        
+
     out = ((x - minval) / maxmin)
     return out, minval, maxmin
 
 res_secp_da, minval_rp, maxmin_rp = normalize(res_secp_da)
 res_true_da, minval_rt, maxmin_rt = normalize(res_true_da)
 
-da_HR, minval_HR, maxmin_HR = normalize(da_HR)
+da_HR, minval_HR, maxmin_HR = normalize(da_HR, 0, 1)
 da_HR_LR_HR, _, _ = normalize(da_HR_LR_HR, minval_HR, maxmin_HR)
 
-# Create simple convolutional network
+class ResBlock(nn.Module):
+    def __init__(self, n_feats = 64,
+                 kernel_size = (3,3),
+                 res_scale = 0.1):
+        super(ResBlock, self).__init__()
+
+        self.n_feats = n_feats
+        self.kernel_size = kernel_size
+        self.res_scale = res_scale
+        
+        self.relu = nn.LeakyReLU(0.01)
+        self.conv = nn.Conv2d(self.n_feats,self.n_feats,stride=(1,1),
+                              kernel_size=self.kernel_size,
+                              padding=1)
+
+        # simple residual block
+        self.resblock = nn.Sequential(self.conv,
+                                      self.relu,
+                                      self.conv)
+
+    def forward(self, x):
+        res = self.resblock(x).mul(self.res_scale)
+        res += x        
+        return res
+
 class ConvNN(nn.Module):
 
     def __init__(self):
@@ -118,37 +144,65 @@ class ConvNN(nn.Module):
 
         self.relu = nn.LeakyReLU(0.01)
         self.kernel_size = (3,3)
-        self.conv1 = nn.Conv2d(1,32,stride=(1,1),
+        self.res_scale = 0.1
+        self.n_blocks = 32
+        self.n_feats = 16
+
+        self.conv1 = nn.Conv2d(1,self.n_feats,stride=(1,1),
                                kernel_size=self.kernel_size,
                                padding=1)
 
-        self.conv2 = nn.Conv2d(32,64,stride=(1,1),
+        self.conv2 = nn.Conv2d(self.n_feats,self.n_feats,stride=(1,1),
                                kernel_size=self.kernel_size,
                                padding=1)
 
-        self.conv3 = nn.Conv2d(64,64,stride=(1,1),
+        self.conv3 = nn.Conv2d(self.n_feats,1,stride=(1,1),
                                kernel_size=self.kernel_size,
                                padding=1)
 
-        self.convT1 = nn.ConvTranspose2d(64,32,stride=(1,1),
+        self.convT2 = nn.ConvTranspose2d(self.n_feats,self.n_feats,stride=(1,1),
                                          kernel_size=self.kernel_size,
                                          padding=1)
 
-        self.convT2 = nn.ConvTranspose2d(32,1, stride=(1,1),
+        self.convT1 = nn.ConvTranspose2d(self.n_feats,1, stride=(1,1),
                                          kernel_size=self.kernel_size,
                                          padding=1)
         self.sigm = nn.Sigmoid()
 
+        blocks = [ ResBlock(self.n_feats, self.kernel_size, self.res_scale)
+                   for _ in range(self.n_blocks) ]
+        blocks.append(self.conv2)
+        self.residual = nn.Sequential(*blocks)
+
+        uo_minval = -1.71300008
+        uo_maxmin = 2.86700014
+
+        scaling = 'minmax1'
+        if scaling == 'minmax2':
+            uo_scale = 2 / uo_maxmin
+            uo_shift = uo_minval + 1 / uo_scale
+        elif scaling == 'minmax1':
+            uo_scale = 1 / uo_maxmin
+            uo_shift = uo_minval
+
+        self.shift_val = uo_shift
+        self.scale_val = uo_scale
+
+    def scale(self, x):
+        x = (x - self.shift_val) * self.scale_val
+        return x
+
+    def unscale(self, x):
+        x = x / self.scale_val + self.shift_val
+        return x
 
     def forward(self, x):
+        x = self.scale(x)        
         x = self.conv1(x)
-        x = self.relu(x)
-        x = self.conv2(x)
-        x = self.relu(x)
-        x = self.convT1(x)
-        x = self.relu(x)
-        x = self.convT2(x)
-        x = self.sigm(x)
+        res = self.residual(x)
+        res += x
+        x = self.convT1(res)
+        x = self.unscale(x)
         return x
 
 model = ConvNN()
@@ -191,7 +245,7 @@ def to_tensor(xb):
     return x
 
 # construct the optimizer
-learning_rate = 0.0005
+learning_rate = 0.0001
 optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
 
 # define loss function
@@ -202,10 +256,10 @@ start_time = time.time()
 log_dict = {'training loss per batch' : [],
             'training loss per epoch' : []}
 
-epochs = 1
+epochs = 5
 verbosity = 20
-do_training=True
-# model.load_state_dict(torch.load('ConvNN.pth'))
+do_training=False
+model.load_state_dict(torch.load('ConvNN.pth'))
 if do_training:
     for epoch_i in range(epochs):
 
@@ -221,7 +275,7 @@ if do_training:
             optimizer.zero_grad()
 
             # compute the loss, model output: o, truth: y
-            loss = loss_fn(o, y, reduction='sum')
+            loss = loss_fn(o, y)
 
             # compute gradients
             loss.backward()
@@ -237,65 +291,46 @@ if do_training:
                       f' Batch {batch_i}/{len(xb_train)} |'
                       f' Loss {loss}')
         print(f'time elapsed: {(time.time()-start_time)/60}m')
-    print(f'total time elapsed: {(time.time()-start_time)/60}m')
+    print(f'total time elapsed: {(time.time()-start_time)/60}m')    
     torch.save(model.state_dict(), 'ConvNN.pth')
 else:
     model.load_state_dict(torch.load('ConvNN.pth'))
 
+pred_steps = len(test_range)
 
-
-for batch_i, (xb, yb) in enumerate(zip(xb_test, yb_test)):
-    x = to_tensor(xb)
-    y = to_tensor(yb)
-    o = model(x)
-    # compute the loss, model output: o, truth: y
-    loss = loss_fn(o, y, reduction='sum')
-    print(loss)
-    break
-    
-    
 with torch.set_grad_enabled(False):
     model.eval()
 
-    init_idx = train_range[-1]
-    xk_HR = da_HR[init_idx, :, :].values
-    xk_HR_m = da_HR[init_idx-1, :, :].values
-
     predY = xr.zeros_like(da_HR[test_range,:,:])
 
-    for i in range(4*24*4):
-
-        xk_LR = da_HR_LR_HR[init_idx+i, :, :].values
+    for i in range(pred_steps):
         xk_LR_p = da_HR_LR_HR[init_idx+i+1, :, :].values
         xk_HR_p = da_HR[init_idx+i+1, :, :].values
 
-        res_secp = sec_pred(xk_HR, xk_HR_m) - xk_LR_p
-        res_true = xk_HR_p - xk_LR_p
-
-        res_secp = (res_secp - minval_rt) / maxmin_rt
-        res_true = (res_true - minval_rt) / maxmin_rt
-
-        x = torch.tensor(np.expand_dims(res_secp.astype('float32'),
+        x = torch.tensor(np.expand_dims(xk_LR_p.astype('float32'),
                                         axis=(0,1)))
-        y = torch.tensor(np.expand_dims(res_true.astype('float32'),
+        y = torch.tensor(np.expand_dims(xk_HR_p.astype('float32'),
                                         axis=(0,1)))
         o = model(x)
-        loss = loss_fn(o, y, reduction='sum')
+        loss = loss_fn(o, y)
         print(loss)
 
-        o_out = o.detach().numpy().squeeze() * maxmin_rt + minval_rt
 
-        xk_HR_m = xk_HR
-        xk_HR = xk_LR_p + o_out
-        predY[i,:,:] = o_out
+        o_out = o.detach().numpy().squeeze() * maxmin_HR + minval_HR
+        xk_LR_p = xk_LR_p * maxmin_HR + minval_HR
+        r = o_out - xk_LR_p
+        predY[i,:,:] = r
 
 
+# EXPORT        
+timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
+    
 model_type='ConvNN'
 pmodel = xr.zeros_like(da_HR[test_range,:,:])
 pmodel[:,:,:] = predY
 
-dsX_HR = da_HR[test_range,:,:]
-dsX_LR = da_HR_LR_HR[test_range,:,:]
+dsX_HR = da_HR[test_range,:,:] * maxmin_HR + minval_HR
+dsX_LR = da_HR_LR_HR[test_range,:,:] * maxmin_HR + minval_HR
 
 dmodel  = dsX_HR - pmodel - dsX_LR
 
@@ -315,7 +350,6 @@ ds['5']['vmin'] = -1/3; ds['5']['vmax'] = 1/3
 ds['6']['vmin'] = -1/3; ds['6']['vmax'] = 1/3
 
 def plot_frame(i):
-
     def plot_wrapper(ds, i):
         ds[i,:,:].plot.pcolormesh(vmin=ds.vmin, vmax=ds.vmax,
                                   center=0, cmap='RdBu',
@@ -334,30 +368,58 @@ def plot_frame(i):
     frame_name = f'output/frame-{i:06d}.png'
     plt.savefig(frame_name)
 
-fig, axs = plt.subplots(2, 2, figsize=(16, 12),
+fig, axs = plt.subplots(2, 2, figsize=(10, 12),
                         sharex=True, sharey=True)
+
 tic = time.time()
 
-with Pool(8) as p:
-    p.map(plot_frame, range(0,4*24*4,1))
+make_movie=False
+if make_movie:
+    with Pool(8) as p:
+        p.map(plot_frame, range(0,pred_steps,1))
 
-from datetime import datetime
-timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
-    
-movie_name = f'movie_{model_type}_{timestamp}.mov'
-framerate = 24
-sys_cmd = ( f"ffmpeg -r {framerate} -f image2 -pattern_type glob -i "
-            f"'output/frame-*.png' "
-            f"-vcodec libx264 -crf 25  -pix_fmt yuv420p -y "
-            f"output/{movie_name}" )
+    movie_name = f'movie_{model_type}_{timestamp}.mov'
+    framerate = 24
+    sys_cmd = ( f"ffmpeg -r {framerate} -f image2 -pattern_type glob -i "
+                f"'output/frame-*.png' "
+                f"-vcodec libx264 -crf 25  -pix_fmt yuv420p -y "
+                f"output/{movie_name}" )
 
-print(sys_cmd)
-os.system(sys_cmd)
-sys_cmd = ( f"rm output/frame-*.png")
-print(sys_cmd)
-os.system(sys_cmd)
+    print(sys_cmd)
+    os.system(sys_cmd)
+    sys_cmd = ( f"rm output/frame-*.png")
+    print(sys_cmd)
+    os.system(sys_cmd)
+
+
+def plot_hovm(mode='horizontal'):
+    plt.close('all')
+    fig, axs = plt.subplots(3, 2, figsize=(10, 12),
+                            sharex=True, sharey=True)
+    for p in range(0,6):
+        plt.subplot(3,2,p+1)
+        breakpoint
+        this_ds = ds[f'{p+1}']
+        this_ds.sel(lat=57.892895).plot.pcolormesh(vmin=this_ds.vmin,
+                                                   vmax=this_ds.vmax,
+                                                   center=0, cmap='RdBu',
+                                                   extend='both',
+                                                   cbar_kwargs={'label':''})
+
+        plt.gca().set_title(this_ds.name)
+        plt.gca().set_ylabel('')
+        plt.gca().set_xlabel('')
+
+
+    frame_name = f'output/hovm_{timestamp}.png'
+    plt.tight_layout()
+    plt.savefig(frame_name)
+
+make_hovm=True
+if make_hovm:
+    plot_hovm()
 
 toc = time.time()
 print(f'elapsed: {toc-tic:02f}')
 print('\a')
-        
+breakpoint()
